Modify how TextParser responds to failed read
On some filehandles eof() does not correctly report the end of the
file. To this end I've changed the code so that confess() is only
called if the _first_ call to read from the filehandle fails.

A fail after one or more successful reads is then treated as
reaching eof(). After reaching eof() in this way, a further call
to read_block() will then trigger an error and confess().