Add VCF4 format for validation
Previoiusly because of the lack of VCF4 format definition, the user upload of VCF
type data (stored in text) could not be validated. This change is intended to add
this missing information so that those VCF records could be validated and parsed.

Also add space as a valid delimiter in addition to tab. I know this is violating
the VCF4.x spece. But I could not get tab delimited VCF record to work. I don't
know why.

Related to ENSEMBL-5050.