Fix ensembl-rest issue 610: close on unblessed ref
A user reported an error for the REST API in
Ensembl/ensembl-rest/issues/610.
This originates in the TabixParser, where we try to close something
that's not our Tabix iterator. This was fixed for another case in sub
seek in the same file a few years ago. The root cause is still unknown.